Run test against a separate database
Apparently there was at some point a ``db`` option to specify a separate
database but this has been deprecated. The depracation warning was
actually raised by the test suite:

    /home/travis/virtualenv/python2.7.9/lib/python2.7/site-packages/DjangoLibrary/__init__.py:87:
    UserWarning: Using the DjangoLibrary 'db' parameter is deprecated. Use
    the 'settings' parameter instead to set a database connection.

Using settings is indeed probably a better way to do this as is
allows to set other separate testing resources, for example a separate
testing web service or search index.

The problem was that, as mentioned in issue #25, this parameter wasn't
taken into account by database manipulation keywords.

So although the initial intention of this commit is to fix #24 it also
fixes #25.